Moved ReduceArray mixin as ArrayUtilities.
